Bug fix for Bug #24: Fix "Best Rectangle Selector"
Christy and I made a few updates to this, as follows:

1) for each rectangle, calculate the perimeter and the area.
2) ensure that the isClosed=True flag is set when calling the cv.ArcLength
function so perimeter is calculated correctly.
3) construct a ratio as follows...
   r = area/(perimeter/4)^2
   ...which normalizes the shape and assigns it a value between 0 and 1, where
perfect squares are 1 and shapes approaching a line (long and skinny) have a
value close to 0.
4) select the polygon whose r is closest to 1.

Haven't done a mathematical proof to ensure that this always works, but
empirically it appears to be working.  Will monitor it...